Fix GCM XMPP client create from hanging
If the client creation fails (not during the first run) then the code will currently hang. This change will prevent this from happening. The reason it is hanging is because nothing is reading from cerr so when an error is being written to the channel it blocks forever. If we make the channel buffered (only size of 1 is necessary) then it will no longer block forever.

Fixes #26